show maintenance on integration page, reuse component within table
